relocated the View Models, created SD utility, added references to projects
