+++ RuggedBooks/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using RuggedBooks
-@using RuggedBooks.Models
+@using RuggedBooksModels.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
