+++ RuggedBooksDAL/RuggedBooksDAL.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -10,6 +10,11 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.4" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\RuggedBooksModels\RuggedBooksModels.csproj" />
+    <ProjectReference Include="..\RuggedBooksUtilities\RuggedBooksUtilities.csproj" />
   </ItemGroup>
 
 </Project>
